Split settings control parts
